+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - The `serde` feature, enabling serialization of `Entity` handles, and a `serialization` module to
   simplify (de)serializing worlds
 - `World::len()` exposing the number of live entities
+- `ColumnBatch` for efficiently spawning collections of entities with the same components when those
+  components' types are not statically known
 
 # 0.3.1 (November 9, 2020)